requirements-svc.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

# Совместимые с Windows 10 и TF CPU версии
numpy==1.24.4
Pillow==10.3.0
opencv-python==4.8.1.78
tensorflow-cpu==2.10.1
# keras подтянется из tensorflow
pydicom==2.3.1
nibabel==5.2.1
rarfile==4.2


